Corregir migraciones.

+++ database/migrations/2024_02_20_154544_create_equipos_table.php
@@ -20,7 +20,7 @@ return new class extends Migration
             $table->string('modelo')->nullable();
             $table->text('descripcion')->nullable();
             $table->tinyInteger('baja')->default(0);
-            $table->integer('aula_id');
+            $table->unsignedBigInteger('aula_id');
             $table->integer('puesto')->nullable();
             $table->timestamps();
